+++ src/main/java/com/example/bcsd/model/Board.java
@@ -1,19 +1,19 @@
 package com.example.bcsd.model;
 
 public class Board {
-    private int id;
+    private Long id;
     private String title;
 
-    public Board(int id, String title) {
+    public Board(Long id, String title) {
         this.id = id;
         this.title = title;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
